test: cover map explorer place links
